requirements.txt
PyYAML
coverage
nose
jinja2
Sphinx
tornado
Werkzeug
pyflaskes